# Replace expensive isinstance checks with faster alternatives