--- a/CHANGES/11085.misc.rst
+++ b/CHANGES/11085.misc.rst
@@ -1,0 +1,1 @@
+Improved performance of isinstance checks by using collections.abc types instead of typing module equivalents -- by :user:`bdraco`.