<!-- Thank you for your contribution! -->

## What do these changes do?

## Summary

This PR optimizes performance by replacing `isinstance()` checks that use types from the `typing` module with their equivalents from `collections.abc`. This avoids expensive runtime type checking overhead while maintaining full backward compatibility.

## Performance Impact

The `typing` module's generic types perform expensive `__instancecheck__` and `__subclasscheck__` operations at runtime. By using `collections.abc` types instead, we achieve significant performance improvements:

- **Mapping checks**: 2.14x faster
- **Sequence checks**: 2.08x faster  
- **Iterable checks**: 2.07x faster
- **Average speedup**: 2.10x

## Changes

### Modified Files

1. **aiohttp/cookiejar.py**
   - Added import: `from collections.abc import Mapping as ABCMapping`
   - Replaced `isinstance(cookies, Mapping)` with `isinstance(cookies, ABCMapping)` at line 241

2. **aiohttp/client_reqrep.py**
   - Added import: `from collections.abc import Mapping as ABCMapping`
   - Replaced `isinstance(cookies, Mapping)` with `isinstance(cookies, ABCMapping)` at line 1010

3. **aiohttp/multipart.py**
   - Added import: `from collections.abc import Mapping as ABCMapping, Sequence as ABCSequence`
   - Replaced `isinstance(obj, (Sequence, Mapping))` with `isinstance(obj, (ABCSequence, ABCMapping))` at line 961
   - Replaced `isinstance(obj, Mapping)` with `isinstance(obj, ABCMapping)` at line 966

4. **aiohttp/payload.py**
   - Added import: `from collections.abc import Iterable as ABCIterable`
   - Replaced `isinstance(type, Iterable)` with `isinstance(type, ABCIterable)` at line 141

## Technical Details

The optimization targets hot paths in the codebase where isinstance checks are performed frequently:
- Cookie handling in requests and responses
- Multipart form data processing  
- Payload type registration

These isinstance checks previously triggered the typing module's complex runtime type checking machinery. The collections.abc types provide the same interface contracts but with native C-level performance.

## Backward Compatibility

This change is fully backward compatible. The typing module types and collections.abc types represent the same interfaces - the only difference is runtime performance. All existing code will continue to work exactly as before.

## Testing

All existing tests pass without modification, confirming that this is a pure performance optimization with no behavioral changes.

## Benchmark Script

A benchmark script is included (`benchmark_isinstance_types.py`) that demonstrates the performance improvements:

```bash
python benchmark_isinstance_types.py
```

Example output:
```
============================================================
isinstance() Performance Benchmark
Comparing typing module vs collections.abc
============================================================

Mapping check (used in cookiejar.py, client_reqrep.py)
--------------------------------------------------
typing.Mapping        0.1652 seconds
abc.Mapping           0.0774 seconds
Speedup: 2.14x faster with collections.abc

Sequence check (used in multipart.py)
--------------------------------------------------
typing.Sequence       0.1635 seconds
abc.Sequence          0.0786 seconds
Speedup: 2.08x faster with collections.abc

Iterable check (used in payload.py)
--------------------------------------------------
typing.Iterable       0.1631 seconds
abc.Iterable          0.0788 seconds
Speedup: 2.07x faster with collections.abc

Simulating CookieJar.update_cookies() hot path
--------------------------------------------------
Old approach (typing):        0.0165 seconds
Optimized (collections.abc):  0.0077 seconds
Speedup: 2.14x faster

============================================================
SUMMARY: Average speedup = 2.10x
============================================================
```


## Are there changes in behavior for the user?

no

## Is it a substantial burden for the maintainers to support this?