@bdraco 2025-05-29
```
#!/usr/bin/env python3
"""
Benchmark script demonstrating the performance difference between
isinstance() checks using typing module types vs collections.abc types.

This script shows why replacing typing.Mapping, typing.Sequence, etc.
with their collections.abc equivalents provides significant performance gains.
"""

import time
from collections.abc import Mapping as ABCMapping, Sequence as ABCSequence, Iterable as ABCIterable
from typing import Mapping as TypingMapping, Sequence as TypingSequence, Iterable as TypingIterable


def benchmark_isinstance(description, test_object, typing_type, abc_type, iterations=1_000_000):
    """Benchmark isinstance checks for a specific type."""
    print(f"\n{description}")
    print("-" * 50)
    
    # Benchmark typing module isinstance
    start = time.perf_counter()
    for _ in range(iterations):
        isinstance(test_object, typing_type)
    typing_duration = time.perf_counter() - start
    
    # Benchmark collections.abc isinstance
    start = time.perf_counter()
    for _ in range(iterations):
        isinstance(test_object, abc_type)
    abc_duration = time.perf_counter() - start
    
    # Calculate speedup
    speedup = typing_duration / abc_duration
    
    # Print results
    print(f"typing.{typing_type.__name__:<12} {typing_duration:>8.4f} seconds")
    print(f"abc.{abc_type.__name__:<16} {abc_duration:>8.4f} seconds")
    print(f"Speedup: {speedup:.2f}x faster with collections.abc")
    
    return speedup


def simulate_cookie_update():
    """Simulate the hot path in CookieJar.update_cookies()"""
    cookies = {"session": "abc123", "user": "john", "token": "xyz789"}
    iterations = 100_000
    
    print("\nSimulating CookieJar.update_cookies() hot path")
    print("-" * 50)
    
    # Using typing.Mapping (old approach)
    start = time.perf_counter()
    for _ in range(iterations):
        if isinstance(cookies, TypingMapping):
            items = cookies.items()
    typing_duration = time.perf_counter() - start
    
    # Using collections.abc.Mapping (optimized approach)
    start = time.perf_counter()
    for _ in range(iterations):
        if isinstance(cookies, ABCMapping):
            items = cookies.items()
    abc_duration = time.perf_counter() - start
    
    speedup = typing_duration / abc_duration
    print(f"Old approach (typing):        {typing_duration:.4f} seconds")
    print(f"Optimized (collections.abc):  {abc_duration:.4f} seconds")
    print(f"Speedup: {speedup:.2f}x faster")


def main():
    print("=" * 60)
    print("isinstance() Performance Benchmark")
    print("Comparing typing module vs collections.abc")
    print("=" * 60)
    
    # Test objects
    test_dict = {"key": "value", "another": "item"}
    test_list = [1, 2, 3, 4, 5]
    test_tuple = (10, 20, 30, 40, 50)
    
    speedups = []
    
    # Benchmark different type checks
    speedups.append(
        benchmark_isinstance(
            "Mapping check (used in cookiejar.py, client_reqrep.py)",
            test_dict, TypingMapping, ABCMapping
        )
    )
    
    speedups.append(
        benchmark_isinstance(
            "Sequence check (used in multipart.py)",
            test_list, TypingSequence, ABCSequence
        )
    )
    
    speedups.append(
        benchmark_isinstance(
            "Iterable check (used in payload.py)",
            test_list, TypingIterable, ABCIterable
        )
    )
    
    # Show real-world simulation
    simulate_cookie_update()
    
    # Summary
    avg_speedup = sum(speedups) / len(speedups)
    print("\n" + "=" * 60)
    print(f"SUMMARY: Average speedup = {avg_speedup:.2f}x")
    print("=" * 60)
    print("\nThis optimization improves performance in hot paths like:")
    print("- Cookie processing in every HTTP request/response")
    print("- Multipart form data handling")
    print("- Payload type registration and processing")
    print("\nThe performance gain is achieved by avoiding the expensive")
    print("__instancecheck__ machinery in the typing module.")


if __name__ == "__main__":
    main()
```